None of these if statementes are necessary anymore
Because of the way orange did %durability... it was weird, but these are no longer necessary. Also, why do you put amtIron -= 1.0D at the end of only diamond and gold? If you want to give the user 1 less ingot for any salvage, that will send them the wrong error message if they do not have enough durability left to get a single ingot (as the if statements I deleted were <1, not <=1).